requirements.txt
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
